Reorganized libs
